Refactor MovieHorizontalItem
