+++ androidHelpers/src/main/java/com/example/androidHelpers/compose/views/Texts.kt
@@ -3,9 +3,15 @@ package com.example.androidHelpers.compose.views
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
+import com.example.androidHelpers.compose.styles.getHeaderTextStyle
 import com.example.androidHelpers.compose.styles.getTitleTextStyle
 
 @Composable
 fun TextTitle(modifier: Modifier = Modifier, text: String) {
     Text(modifier = modifier, text = text, style = getTitleTextStyle())
+}
+
+@Composable
+fun TextHeader(modifier: Modifier = Modifier, text: String) {
+    Text(modifier = modifier, text = text, style = getHeaderTextStyle())
 }
